test: move proxy tests into proxies.spec.js
Consolidate the proxy fetch and dispatcher tests out of client.spec.js
and webhook.spec.js into a dedicated proxies.spec.js that matches the
src/proxies.js module.
